+++ rpc-server.ts
@@ -20,7 +20,8 @@ import { Logger, loggers } from "winston";
 
 import { buildSchemaValidators } from "./rpc/validators";
 import { loadContractFixtures } from "./rpc/contracts";
-import { RPCLoad } from "./rpc/load";
+import { RPCVault } from "./rpc/vault";
+import { RPCLoad as RPCLoad_1_0_2 } from "./rpc/Load/1.0.2/RPCLoad";
 import { RPCEvent, startEventSubscriptions } from "./rpc/event";
 import { RPCWallet, setupWalletEncryptionHandler } from "./rpc/wallet";
 import { RPCTransaction } from "./rpc/transaction";
@@ -75,34 +76,40 @@ server.expose("transaction", {
 });
 
 server.expose("load", {
-    "create_vault": RPCLoad.CreateVault,
-    "create_shipment_transaction": RPCLoad.CreateShipmentTx,
-    "update_vault_hash_transaction": RPCLoad.UpdateVaultHashTx,
-    "fund_eth_transaction": RPCLoad.FundEthTx,
-    "fund_cash_transaction": RPCLoad.FundCashTx,
-    "fund_ship_transaction": RPCLoad.FundShipTx,
-    "commit_to_shipment_transaction": RPCLoad.CommitToShipmentTx,
-    "shipment_in_transit_transaction": RPCLoad.ShipmentInTransitTx,
-    "carrier_complete_transaction": RPCLoad.CarrierCompleteTx,
-    "shipper_accept_transaction": RPCLoad.ShipperAcceptTx,
-    "shipper_cancel_transaction": RPCLoad.ShipperCancelTx,
-    "pay_out_transaction": RPCLoad.PayOutTx,
-    "get_shipment_details": RPCLoad.GetShipmentDetails,
-    "get_shipment_details_continued": RPCLoad.GetShipmentDetailsContinued,
-    "get_escrow_status": RPCLoad.GetEscrowStatus,
-    "get_contract_flags": RPCLoad.GetContractFlags,
-    "get_tracking_data": RPCLoad.GetTrackingData,
-    "add_tracking_data": RPCLoad.AddTrackingData,
-    "get_shipment_data": RPCLoad.GetShipmentData,
-    "add_shipment_data": RPCLoad.AddShipmentData,
-    "get_document": RPCLoad.GetDocument,
-    "add_document": RPCLoad.AddDocument,
-    "list_documents": RPCLoad.ListDocuments,
-    "verify_vault": RPCLoad.VerifyVault,
+    "create_shipment_transaction": RPCLoad_1_0_2.CreateShipmentTx,
+});
 
-    "get_historical_shipment_data": RPCLoad.GetHistoricalShipmentData,
-    "get_historical_tracking_data": RPCLoad.GetHistoricalTrackingData,
-    "get_historical_document": RPCLoad.GetHistoricalDocument,
+server.expose("load-1.0.2", {
+    "create_shipment_transaction": RPCLoad_1_0_2.CreateShipmentTx,
+    "update_vault_hash_transaction": RPCLoad_1_0_2.UpdateVaultHashTx,
+    "fund_eth_transaction": RPCLoad_1_0_2.FundEthTx,
+    "fund_cash_transaction": RPCLoad_1_0_2.FundCashTx,
+    "fund_ship_transaction": RPCLoad_1_0_2.FundShipTx,
+    "commit_to_shipment_transaction": RPCLoad_1_0_2.CommitToShipmentTx,
+    "shipment_in_transit_transaction": RPCLoad_1_0_2.ShipmentInTransitTx,
+    "carrier_complete_transaction": RPCLoad_1_0_2.CarrierCompleteTx,
+    "shipper_accept_transaction": RPCLoad_1_0_2.ShipperAcceptTx,
+    "shipper_cancel_transaction": RPCLoad_1_0_2.ShipperCancelTx,
+    "pay_out_transaction": RPCLoad_1_0_2.PayOutTx,
+    "get_shipment_details": RPCLoad_1_0_2.GetShipmentDetails,
+    "get_shipment_details_continued": RPCLoad_1_0_2.GetShipmentDetailsContinued,
+    "get_escrow_status": RPCLoad_1_0_2.GetEscrowStatus,
+    "get_contract_flags": RPCLoad_1_0_2.GetContractFlags,
+});
+
+server.expose("vault", {
+    "create_vault": RPCVault.CreateVault,
+    "get_tracking_data": RPCVault.GetTrackingData,
+    "add_tracking_data": RPCVault.AddTrackingData,
+    "get_shipment_data": RPCVault.GetShipmentData,
+    "add_shipment_data": RPCVault.AddShipmentData,
+    "get_document": RPCVault.GetDocument,
+    "add_document": RPCVault.AddDocument,
+    "list_documents": RPCVault.ListDocuments,
+    "verify_vault": RPCVault.VerifyVault,
+    "get_historical_shipment_data": RPCVault.GetHistoricalShipmentData,
+    "get_historical_tracking_data": RPCVault.GetHistoricalTrackingData,
+    "get_historical_document": RPCVault.GetHistoricalDocument,
 });
 
 server.expose("event", {
